+++ fractal_server/runner/executors/local/runner.py
@@ -22,9 +22,7 @@ from fractal_server.runner.executors.call_command_wrapper import (
 )
 from fractal_server.runner.task_files import TaskFiles
 from fractal_server.runner.v2.db_tools import bulk_update_status_of_history_unit
-from fractal_server.runner.v2.db_tools import (
-    update_status_of_history_unit_no_commit,
-)
+from fractal_server.runner.v2.db_tools import update_history_unit_no_commit
 
 logger = set_logger(__name__)
 
@@ -159,7 +157,7 @@ class LocalRunner(BaseRunner):
             result = None
             exception = TaskExecutionError(str(e))
             with next(get_sync_db()) as db:
-                update_status_of_history_unit_no_commit(
+                update_history_unit_no_commit(
                     history_unit_id=history_unit_id,
                     status=HistoryUnitStatus.FAILED,
                     db_sync=db,
@@ -168,6 +166,7 @@ class LocalRunner(BaseRunner):
                 return None, exception
 
         # RETRIEVAL PHASE
+
         with next(get_sync_db()) as db:
             try:
                 result = future.result()
@@ -176,7 +175,7 @@ class LocalRunner(BaseRunner):
                     TaskType.COMPOUND,
                     TaskType.CONVERTER_COMPOUND,
                 ]:
-                    update_status_of_history_unit_no_commit(
+                    update_history_unit_no_commit(
                         history_unit_id=history_unit_id,
                         status=HistoryUnitStatus.DONE,
                         db_sync=db,
@@ -184,7 +183,7 @@ class LocalRunner(BaseRunner):
                 return result, None
             except Exception as e:
                 logger.debug("[submit] END with exception")
-                update_status_of_history_unit_no_commit(
+                update_history_unit_no_commit(
                     history_unit_id=history_unit_id,
                     status=HistoryUnitStatus.FAILED,
                     db_sync=db,
@@ -296,7 +295,7 @@ class LocalRunner(BaseRunner):
                     exceptions[positional_index] = TaskExecutionError(str(e))
                     if task_type == TaskType.PARALLEL:
                         with next(get_sync_db()) as db:
-                            update_status_of_history_unit_no_commit(
+                            update_history_unit_no_commit(
                                 history_unit_id=current_history_unit_id,
                                 status=HistoryUnitStatus.FAILED,
                                 db_sync=db,
@@ -322,7 +321,7 @@ class LocalRunner(BaseRunner):
                         try:
                             results[positional_index] = fut.result()
                             if task_type == TaskType.PARALLEL:
-                                update_status_of_history_unit_no_commit(
+                                update_history_unit_no_commit(
                                     history_unit_id=current_history_unit_id,
                                     status=HistoryUnitStatus.DONE,
                                     db_sync=db,
@@ -338,7 +337,7 @@ class LocalRunner(BaseRunner):
                                 str(e)
                             )
                             if task_type == TaskType.PARALLEL:
-                                update_status_of_history_unit_no_commit(
+                                update_history_unit_no_commit(
                                     history_unit_id=current_history_unit_id,
                                     status=HistoryUnitStatus.FAILED,
                                     db_sync=db,